Code from the Challenge Scenario
The last part of this session was to attempt and replicate a website that Eleni built that would allow the student to test their knowledge on basic HTMl/CSS topics discussed in the live demo. All that was provided was some hints and an image of what the finished website looked like.